+++ flask_fs/backends/s3.py
@@ -86,14 +86,15 @@ class S3Backend(BaseBackend):
         for f in self.bucket.objects.all():
             yield f.key
 
-    def metadata(self, filename):
+    def get_metadata(self, filename):
         '''Fetch all availabe metadata'''
         obj = self.bucket.Object(filename)
         checksum = 'md5:{0}'.format(obj.e_tag[1:-1])
+        mime = obj.content_type.split(';', 1)[0] if obj.content_type else None
         return {
             'checksum': checksum,
             'size': obj.content_length,
-            'mime': obj.content_type,
+            'mime': mime,
             'modified': obj.last_modified,
         }