+++ flask_fs/__about__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-__version__ = '0.5.2.dev'
+__version__ = '0.6.0'
 __description__ = 'Simple and easy file storages for Flask'