+++ CHANGELOG.rst
@@ -5,8 +5,9 @@ Current
 -------
 
 - Added ``copy()`` and ``move()`` operations
-- `delete()` now supports directories (or prefixes for key/value stores)
-- Improve `metadata()` `mime` handling
+- ``delete()`` now supports directories (or prefixes for key/value stores)
+- Improve ``metadata()`` ``mime`` handling
+- Added explicit ``ImageField.full(external=False)``
 
 0.5.1 (2018-03-12)
 ------------------